pyproject.toml
[tool.poetry]
name = "nthp_api"
version = "0.1.0"
description = ""
authors = ["Will Pimblett <will@wjdp.uk>"]
license = "MIT"

[tool.poetry.scripts]
nthp = "nthp_api.cli:cli"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.7"
python-frontmatter = "^1.0.0"
peewee = "^3.16.2"
python-slugify = "^8.0.1"
pyhumps = "^3.8.0"
click = "^8.1.3"
coloredlogs = "^15.0.1"
pydantic-collections = "^0.4.0"
Markdown = "^3.4.3"
PyYAML = "^6.0"
httpx = "^0.24.1"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
pytest-vcr = "^1.0.2"
mypy = "^1.3.0"
pytest-depends = "^1.0.1"
types-python-slugify = "^8"
types-Markdown = "^3.3.7"
freezegun = "^1.2.2"
types-PyYAML = "^6.0.1"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
files = "nthp_api"
pretty = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.ruff]
select = [
    "E", # pycodestyle errors
    "F", # pyflakes
    "I", # isort
    "C90", # mccabe, complexity
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    # "DTZ", # flake8-datetimez
    "ICN", # flake8-import-conventions
    "ISC", # flake8-implicit-str-concat
    "INP", # flake8-no-pep420
    "PIE", # flake8-pie
    "T20", # flake8-print
    "Q", # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SLF", # flake8-self
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "PL", # pylint
    "TRY", # tryceratops
    "RUF", # ruff special rules
]
ignore = [
    "TRY003", # Permit long exception messages
]

# Higher than black to give wiggle room
line-length = 120

[tool.ruff.per-file-ignores]
"bin/*" = ["INP001"]
"dev/*" = ["INP001"]
"*/fixtures/*" = ["E501"]
